requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cohere==4.18.0
cssselect==1.2.0
dataclasses-json==0.5.13
exceptiongroup==1.1.2
fastavro==1.7.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
jieba3k==0.35.1
joblib==1.3.1
langchain==0.0.247
langsmith==0.0.15
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
newspaper3k==0.2.8
nlpcloud==1.1.43
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pydantic==1.10.12
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
selenium==4.10.0
selenium-stealth==1.0.6
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tenacity==8.2.2
tiktoken==0.4.0
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.16.2
gspread==3.7.0
oauth2client==4.1.3
pydrive==1.3.1
google-cloud-secret-manager==2.3.0
